+++ index.html
@@ -4,11 +4,11 @@
         <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <link rel="stylesheet" href="./index.css" />
-        <title>Episode - 1</title>
+        <title>Food Delivery App</title>
     </head>
     <body>
         <div id="root"></div>
         
-        <script type="module" src="./App.js"></script>
+        <script type="module" src="./src/App.js"></script>
     </body>
 </html>
