+++ src/App.js
@@ -1,0 +1,20 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+
+import Header from "./components/Header";
+import Body from "./components/Body";
+import { Footer } from "./components/Footer"; // it is mentioned as name export in Footer.js file
+
+const App = () => {
+    return (
+        <div className="app">
+            <Header />
+            <Body />
+            <Footer />
+        </div>
+    );
+}
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+root.render(<App />);
