+++ src/components/Footer.js
@@ -1,0 +1,12 @@
+export const Footer = () => {
+    return (
+        <div className="footer-ctr">
+            <ul>
+                <li>Copyrights to Musham</li>
+            </ul>
+        </div>
+    )
+}
+
+//export default Footer;
+// I am exporting footer component as named component here
